[feat] Add password encryption when storing sensitive data
Closes #15.
When starting the terminal for the first time the user will be prompt to
create a password, this password is then used to encrypt all sensitive
data, ie private-keys/APIs. The password will always be prompted at the
start of the terminal and will live in memory for the duration of the
program.