[feat] Improve news widget layout.
Re worked how layout is constructed to only include widgets with data.
This ensures we don't have reserved empty spaces on the widet.

Added a hack on the show event to re-calculate minimum heights of labels
using word wrap and rich text. This is a know bug of QT where the height
of this labels can only be calculated after the layouts are resolved.
https://stackoverflow.com/questions/14238138/heightforwidth-label